+++ packages/orm/src/index.ts
@@ -2,4 +2,4 @@ export * from './table';
 export * from './orm';
 export * from './repo';
 export * from './codegen/appCodeUpdater';
-export { columnTypes } from 'pqb';
+export { columnTypes, OrchidOrmError, OrchidOrmInternalError } from 'pqb';