+++ packages/qb/pqb/src/errors.test.ts
@@ -44,6 +44,7 @@ describe('errors', () => {
       }
     }
 
+    expect(err?.query).toBe(UniqueTable);
     expect(err?.isUnique).toBe(true);
     expect(err?.columns).toEqual({
       one: true,
@@ -73,6 +74,7 @@ describe('errors', () => {
       }
     }
 
+    expect(err?.query).toBe(UniqueTable);
     expect(err?.isUnique).toBe(true);
     expect(err?.columns).toEqual({
       thirdColumn: true,